Refactor tile rendering into dedicated components
