+++ src/components/admin/tiles/RichTextEditor.tsx
@@ -1,0 +1,131 @@
+import React, { useEffect } from 'react';
+import { EditorContent, useEditor, type Editor } from '@tiptap/react';
+import StarterKit from '@tiptap/starter-kit';
+import Underline from '@tiptap/extension-underline';
+import { TextStyle } from '@tiptap/extension-text-style';
+import Color from '@tiptap/extension-color';
+import FontFamily from '@tiptap/extension-font-family';
+import BulletList from '@tiptap/extension-bullet-list';
+import OrderedList from '@tiptap/extension-ordered-list';
+import ListItem from '@tiptap/extension-list-item';
+import { TextAlign } from '../../../extensions/TextAlign';
+import FontSize from '../../../extensions/FontSize';
+import type { LessonTile, TextTile } from '../../../types/lessonEditor';
+
+interface RichTextEditorProps {
+  textTile: TextTile;
+  tileId: string;
+  onUpdateTile: (tileId: string, updates: Partial<LessonTile>) => void;
+  onFinishTextEditing: () => void;
+  onEditorReady: (editor: Editor | null) => void;
+  textColor?: string;
+}
+
+export const RichTextEditor: React.FC<RichTextEditorProps> = ({
+  textTile,
+  tileId,
+  onUpdateTile,
+  onFinishTextEditing,
+  onEditorReady,
+  textColor,
+}) => {
+  const editor = useEditor({
+    extensions: [
+      StarterKit.configure({
+        bulletList: false,
+        orderedList: false,
+        listItem: false,
+      }),
+      BulletList.configure({
+        HTMLAttributes: { class: 'bullet-list' },
+        keepMarks: true,
+        keepAttributes: true,
+      }),
+      OrderedList.configure({
+        HTMLAttributes: { class: 'ordered-list' },
+        keepMarks: true,
+        keepAttributes: true,
+      }),
+      ListItem,
+      Underline,
+      TextStyle,
+      Color.configure({ types: ['textStyle'] }),
+      FontFamily.configure({ types: ['textStyle'] }),
+      FontSize,
+      TextAlign.configure({ types: ['paragraph'] }),
+    ],
+    content:
+      textTile.content.richText ||
+      `<p style="margin: 0;">${textTile.content.text || ''}</p>`,
+    onUpdate: ({ editor }) => {
+      const html = editor.getHTML();
+      const plain = editor.getText();
+      onUpdateTile(tileId, {
+        content: {
+          ...textTile.content,
+          text: plain,
+          richText: html,
+        },
+      });
+    },
+    autofocus: true,
+  });
+
+  useEffect(() => {
+    onEditorReady(editor);
+    return () => onEditorReady(null);
+  }, [editor, onEditorReady]);
+
+  if (!editor) return null;
+
+  const handleBlur = (e: React.FocusEvent) => {
+    const toolbar = document.querySelector('.top-toolbar');
+    if (toolbar && e.relatedTarget && toolbar.contains(e.relatedTarget as Node)) {
+      e.preventDefault();
+      editor.commands.focus();
+      return;
+    }
+    onFinishTextEditing();
+  };
+
+  const handleKeyDown = (e: React.KeyboardEvent) => {
+    if (e.key === 'Tab') {
+      e.preventDefault();
+      if (editor.isActive('listItem')) {
+        if (e.shiftKey) {
+          editor.chain().focus().liftListItem('listItem').run();
+        } else {
+          editor.chain().focus().sinkListItem('listItem').run();
+        }
+      } else {
+        editor.chain().focus().insertContent('\t').run();
+      }
+    }
+  };
+
+  return (
+    <div
+      className="w-full h-full p-3 overflow-hidden relative tile-text-content tiptap-editor"
+      style={{
+        fontSize: `${textTile.content.fontSize}px`,
+        fontFamily: textTile.content.fontFamily,
+        color: textColor,
+        display: 'flex',
+        flexDirection: 'column',
+        justifyContent:
+          textTile.content.verticalAlign === 'center'
+            ? 'center'
+            : textTile.content.verticalAlign === 'bottom'
+            ? 'flex-end'
+            : 'flex-start',
+      }}
+    >
+      <EditorContent
+        editor={editor}
+        className="w-full focus:outline-none break-words rich-text-content tile-formatted-text"
+        onBlur={handleBlur}
+        onKeyDown={handleKeyDown}
+      />
+    </div>
+  );
+};
